[ spls-03-2024 ] Talk about ATM, dpairs, and `arb`
There's lots to cover. "Spectrum"-slide will probably make an
appearance, "ergonomics" is tenuous at this stage. It is going to be a
more theory-focused talk than the generic-PhD-audience one. And I need
to remember that the topic is the faults in ISMs and how to spot them
using QC, not _all of my PhD_  ^^;;